Fix orthogonalisation in presence of degeneracies
The code tacitly assumed that at most one eigenvalue degeneracy was
present.  If this was not the case, then `find_duplicates()` would
return several entries, which would subsequently be interpreted as array
indices which were too large for the shape of the indexee.

This adds a loop over all degenerate sets, makes the code a little more
readable and adds documentation to `find_duplicates()` to explain what
it is meant to do.