Allow more flexibility in Hamiltonian functions
Change the call structure of `hamiltonian()` and `dhamiltonian()` in
`floq.System`s to be the much more general
`hamiltonian(*args, **kwargs)`.  Arguments can now be supplied as
additional positional and keyword arguments to all the interesting
methods of a `floq.System`, which are then propagated through.

This has some ramifications for our ability to cache correctly - now
that the call structure is arbitrary, it is difficult to make a full
deep comparison of the arguments (though making a deep copy would have
been easy), so I no longer do that.  The class help message states that
caching will be broken if arguments are mutated in-place, but the actual
implementation should be a little more resilient - each argument and
keyword argument is checked for array equality with its counterpart in
the cache, but this may fail if arrays are passed nested deep inside
another structure.