Allow Hamiltonian and derivatives to be static
The `floq.System` constructor now accepts Hamiltonians and derivatives
which do not have any controls.  If the input to either is not callable,
then internally it will be converted into a callable function which
simply eats arguments passed to it.  This should reduce a little of the
boiler-plate necessary to pass derivatives which are independent of
controls, but also allow completely static Hamiltonians.