Stop ignoring previous value of nz
While changing something around I had removed a line which set
`System.nz` to the last successful value used in the `FixedSystem`
calculations.  This makes `System.nz` a property, so that it always
pulls the last successful value, or the set value if that is more
recent.

This brings the speed of array creation back to parity with the old
Python 2.7 version.