Update to newer numba conventions
When `numba` was introduced as a dependency, Marcel wrote replacements
for `numpy.zeros` and `numpy.outer`.  I'm not sure why that was the
case - perhaps at the time `numba` didn't support those functions?  They
are now supported, though, so I removed the helpers.

I also moved from the old style of type inference `numba.autojit()` to
the simpler `numba.jit()`, which now does the same inference.