Rename helpers module to linalg
The functions in this module were all linear algebra functions, so it
makes sense to have them named accordingly.  This removes a couple of
functions which have corresponding `numpy` versions, since anything we
can pass onto a mature library means less technical debt for us.

This also removed some of the `from ..helpers import matrix as mm`-type
stuff with the cryptic two-letter imports.  I'm not a big fan of that
style (though it's pervasive throughout Python, and `numpy` is an
exception), because I think it's harder to follow what the code is
doing.