Remove custom exceptions
`EigenvalueNumberError` was never referenced (and was more of an
internal logic error anyway).  `NZTooLargeError` is just a specific type
of `RuntimeError`, and the general best practice seems to be to just
raise that rather than having our own specific version.  Dealing with it
would require user intervention anyway, so there doesn't seem to be much
harm in this.

If in the future we want them back, they should derive from the
exceptions that I replaced them with, rather than from the base
`Exception`, so we still get the hierarchy.