Tidy up FixedSystem
This changes the names of all the functions and internals of
`FixedSystem` to match the new names in `floq.System`, and removes a lot
of additional functions like `_compute_u()`, which is now just baked
into the underlying `__compute_eigensystem_and_u()`.

The separated functions were likely there only to facilitate the
automated testing of the caching.  I should come up with a way to have
the best of both worlds - having a lot of additional code just creates
more technical debt and should be avoided, but the concept of testing
all the small details underneath is definitely desirable as well.  Here
I opted to remove the additional functions because the parts which
implement the cache are very straightforward (`if self.x is not None`).

In the future, I will hopefully be removing a lot of this code anyway to
make way for a smarter cache, where only the eigensystem is stored so
multiple times can be evaluated: see #1.