Move examples out of package
The `floq.System` examples `rabi` and `spin` have now been moved into a
separate "examples" folder at the top-level of the repository, because
they oughtn't to be part of the core package.  They are still useful for
integration tests, so the tests now import the examples as modules to
test the functionality.

This should simplify some more of the `floq` package structure, and keep
the maintenance cost from becoming too high.

This is the first steps to a proper fixing of #4.