Begin variable name refactoring
I'm trying to move away from cryptic two-letter variable names towards
more descriptive ones, even if that means more typing.  The current
naming scheme is very difficult to reason about deep inside
`core.evolution`, but it doesn't have to be that way.

It's not a big help for everything to be labelled with the same Greek
letters that Marcel uses in his thesis - everyone uses slightly
different ones, and there's a big possibility of confusion with the
Fourier/non-Fourier-transformed and chi/phi splits in the notation.

I haven't yet touched `FixedSystemParameters` or `core.evolution`
because they're going to part of a much larger refactoring shortly.