Fix off-by-two error in calculate_factors
The `factors` array was erroneously being created with a shape of
    (n_parameters, 2*n_zones + 1, dimension, dimension)
                   ^^^^^^^^^^^^^
where the marked axis should have had dimension `2*n_zones - 1` (with a
minus one, not plus one).  Likewise, the factor of `2 * n_zones` later
used in a `floq.linalg.n_to_i` call should have been `2 * n_zones - 1`.

Fortuitously, neither of these actually caused bugs in the output.  The
former was fine because the extra space in the array was never accessed,
and the latter was fine because of a modulo divison by 2 in the `n_to_i`
call which returned the same value for `2n` and `2n - 1`.

On this vein, there is also a "bug" in Marcel's thesis---the text
underneath equation (2.7) states that there are
    dimension * dimension * 2*n_zones
different combined factors per control parameter, but in fact there are
    dimension * dimension * (2*n_zones - 1).
This obviously has no bearing on any of the physics presented in that
section, though.