Rewrite evolution to calculate U at changing times
This commit leaves in the current evolution internals, but the new code
can be used as a complete, feature-for-feature replacement, except now
all of the operators can be calculated at any time without having to
re-diagonlise the Floquet matrix.  The `floq.System` API still uses the
old code, but that will change in the next commit or two.

The general principle is that all of the time-independent parts are
calculated only once, when the matrix is diagonalised.  This includes
the Floquet basis bras at `t = 0`, the eigenvectors and eigenvalues
(quasienergies) of K, and the derivative matrix dK/dcontrols.  The
calculation of this system is typically the most time-consuming
operation by far.

Once this has been calculated, we can calculate new versions of the
operators for changing the time (but keeping the underlying Hamiltonian
the same) very quickly.  For example, using the spin system in
`examples/spins.py` I created a system using `spin(10, 0.87, 1.01)`.
I used 65 Brillouin zones as defined by Marcel (which was sufficient to
give unitary dynamics), and I timed the new functions (_not_ the
diagonlisation) to take
    New functions with a precomputed diagonalisation:
               u():   0.1010(5) ms (1000 loops)
           du_dt():   0.116(3)  ms (1000 loops)
    du_dcontrols(): 235(4)      ms (100 loops)

Using the same set-up, I timed the entirety of the code path for
`floq.System.u()` and the others, including diagonalisation.  This uses
Marcel's code, so is the baseline for measurement.
    Old functions including diagonalisation:
               u():   7.24(10) ms  (2000 loops)
           du_dt():   7.16(8)  ms  (1000 loops)
    du_dcontrols(): 577(1)     ms  (100 loops)

I then did the same thing for the entire code path for the new code, so
calling `evolution.eigensystem()` followed by `evolution.u()`, etc.  For
`u()` and `du_dt()` I did not pass the derivatives of the Hamiltonian to
`eigensystem()` so that the timings do not include additional
calculations that are not part of that code path.  This behaviour will
be available to the user, so it should be a fair test.
    New functions including diagonalisation:
               u():   6.98(7) ms  (2000 loops)
           du_dt():   6.71(5) ms  (2000 loops)
    du_dcontrols(): 297(4)    ms  (100 loops)

The important takeaway is that we pay no time penalty for this new
behaviour, and in fact we actually get a speedup from a little bit of
optimisation of the code.  The `u()` and `du_dt()` full codepaths gained
about a 5% speedup, but the `du_dcontrols()` became very nearly twice as
fast due to major optimisations in how the factors are calculated,
including making sure all paths have full `numba` support.

The optimisation of `assemble_k()` and `assemble_dk()` is still
required, which may give some speed up, particularly for
higher-dimension Hamiltonians, but the main time-sink is currently the
diagonalisation of the `K` matrix (except for `du_dcontrols()` of
course).

All timing was done on the i5-5278U @ 2.90GHz processor in my CDT
Macbook Pro using `ipython` `%timeit`, where the errors are the standard
deviation of 7 runs, each of which had the indicated number of loops.

This is the first half of fixing #1 and #3.