Add new diagonalisation methods
Similar to 5bec9172, this does not remove any code from
`core.evolution`, but adds a new diagonalisation method `diagonalise()`.
This handles both sparse and dense matrices, and takes extra care to
make sure that degeneracies which fall on the edge of the first
Brillouin zone are handled safely (see #5).  If something cannot be
handled safely, then a `RuntimeError` is raised - this is usually a
problem of precision from the sparse matrix eigensystem solver.

Again, we suffer no major speed penalties for these additional safety
checks, and in fact gain a little bit of speedup by organising where
things are sorted, split, transposed and reshaped.  Previously, these
were unnecessarily repeated, and some arrays were recreated in loops
where vectorised indexing coupled with transposition and reshaping would
have been better.

Using all the same proceedures, and the same system as in 5bec9172, I
timed the new code path and the old code path as
         diagonalise(): 5.59(5) ms
    find_eigensystem(): 6.08(6) ms
using 2000 loops for each, and 71 Brillouin zones.

The new code is essentially a fix of #5, but I cannot close the issue
until `floq.System` uses the new code path.