Correct integral_factors calculation
The integral in Marcel's thesis eq. (1.48) says that the integral is
simplified if `j == j'` and `\Delta\mu == 0`, but this fails if there
are degenerate eigenvalues, or in other cases where their difference
matches some other terms.

The correct behaviour is to use the limited form if the exponent is ever
equal to 0, regardless of what the indices are.