Directly assemble K as a sparse matrix
Previously, the Floquet matrix was built up as a dense matrix, then
`scipy.sparse.csc_matrix()` was called on the result to produce a sparse
matrix.  This did allow us to use the sparse eigenvalue equation solver
to get its speedup, but it wasn't really the most efficient way of doing
things.

It's best to have K as a Compressed Sparse Column format matrix for the
solver, rather than building up a block-form as it most naturally lends
itself to.  The constructor `scipy.sparse.bmat()` is very slow in
comparison to other constructors as well.

We build up the matrix in co-ordinate `ijv` form, which can be
efficiently converted into a `csc` form.

This doesn't offer a big speedup on the full code path because the
timing is dominated by the eigenvalue solver, but I was mostly using
this as practice for subsequently dealing with the `dK` matrix.  In
cases of low sparsity, this method is marginally slower than the
previous one, but the absolute time difference is very small.

For the system in `examples/spins`, using a system created with
`spins(10, 0.87, 1.01)`, I timed the old `K` code by first finding the
Hamiltonian (untimed), then timing
    scipy.sparse.csc_matrix(floq.evolution.assemble_k(h, 65, 1.01))
(so there were 65 Brillouin zones) at first.  I added the call to
`csc_matrix()` to test the whole code path of `K` creation.  For the new
version, I simply timed
    floq.evolution.assemble_k_sparse(h, 65, 1.01)
I did this for 65 Brillouin zones, 251 Brillouin zones, and for a
Molmer--Sorensen-type Hamiltonian (Hamiltonian dimension 60, 121
Brillouin zones)

Results:
             Case                   Old           New
    spins, 65 zones, 14% full:   0.554(8) ms   0.713(9) ms
    spins, 251 zones, 4% full:   3.76(5)  ms   2.08(3)  ms
          MS gate, 0.36% full:   1.05(1)   s   9.9(3)   ms

As the sparsity of the matrix increases, the new way becomes
significantly faster, even if at more full matrices it's a little less
efficient.

This should inform our creation of the `dK` matrix, which is necessarily
at least as sparse as (and generally significantly more empty than) `K`.