Rearrange expectation value to reduce matrix dots
Previously I was doing `dimension` (of the Hamiltonian) vector dot `K`
operations where I only needed to do one.  By rearranging this we can
reduce the number of these we have to do by a lot.

Timing only `evolution.du_dcontrols()` with a precomputed `Eigensystem`
using the classic `spin(10, 0.87, 1.01)` with 65 Brillouin zones, the
old time was 229(2)ms, and the new time is 131(2)ms, which is 43%
faster.

For a Molmer--Sorensen Hamiltonian of dimension 20, 65 Brillouin zones
and one control parameter, the time taken to calculate `du_dcontrols()`
dropped from 88s to 8s.  This is still too slow to really use, but it's
getting closer.

Sparse matrix algebra may help, but we will lose the ability to use
`numba` for those parts.