Convert dK matrix to custom sparse representation
The dK matrix at least as sparse as K, and usually even more so.
Unfortunately, building it as a `scipy.sparse` matrix would have
required losing `numba` support when using it to calculate the factors
of things.  The solution is to implement a custom column-sparse matrix
tuple, and `numba`-compatible necessary operations.

The only necessary operation for each `dK` matrix is
    vector . dK[i],
which motivates the choice of a column-sparse representation.  Since it
transpires that each column in `dK` typically has only 2 or 3 entries
per block, and few populated blocks, this reduces the complexity of
`combined_factors()` from `n_zones^3 * dimension^4` (where `dimension`
is the Hamiltonian dimension) to `n_zones^2 * dimension^3` by reducing
the dot product from `(n_zones*dimension)^2` to simply
`n_zones*dimension`, assuming a constant number of populated elements
per column.

Storing the `dK` matrix in this column-sparse form also has huge
memory-usage ramifications.  The required storage is now linear in the
number of populated elements, plus linear in `n_zones * dimension` (to
store the number of elements for each column).  Since the number of
populated elements is often linear in the dimension of the Hamiltonian,
this means the storage requirements in this case are simply of order
    n_zones * dimension.

The speed-up is very noticeable, and is particularly important as the
sparseness of the problem increases.  Our favourite base-case,
    spin(10, 0.87, 1.01)
from `examples/spins.py` with 65 zones gets more than a 3x speed-up for
the entire code path of `du_dcontrols()` - going from 145(1)ms to
45.8(6)ms.  If we time with a pre-computed eigensystem (so we don't have
to compute `dK`), we go from 134(1)ms to 25.2(5)ms.