Reorder sums in du_dcontrols() for speed
This reduces the time complexity of `evolution.du_dcontrols()` to be of
order
    n_controls * n_zones * dimension^3 + n_controls * dimension^4\
    + n_zones^2 * dimension^3
(the last term comes from `combined_factors()`), down from
    n_controls * (n_zones^2 * dimension^3 + n_zones * dimension^4).

This is likely as far as we can go on asymptotic behaviour---I can't
think of a way to lower the `dimension^4`, and even if we could, the
`n_zones^2 * dimension^3` likely dominates it anyway now, and that's
pretty fixed.

Once again, with everybody's favourite example timing
    spin(10, 0.87, 1.01, n_zones=65)
from `examples/spins.py` (so 20 control parameters!) the full code path
    controls = np.random.rand(20)
    system.u(0.5, c)
    system.du_dcontrols(0.5, c)
now takes a total of 24.2ms on my machine, down from 44.7ms.

This effect is significantly more pronounced for Hamiltonians with a
larger dimension.  My original Molmer--Sorensen Hamiltonian of dimension
60 with 121 Brillouin zones could not calculate `du_dcontrols()` on the
cluster within 24 hours as of commit cb47fe8, but now finishes (with 7
control parameters) in 45 seconds on my laptop (limited by
`combined_factors()`).