Allow more Hamiltonian forms
This allows the Fourier transformed Hamiltonians (and their derivatives)
to be expressed in much more general forms.  This removes the
requirement to have a symmetric distribution of modes, since the
populated modes can now be explicitly specified rather than being
inferred from the shape of the operator array.

Allowing this means that a new canonical form of Fourier transformed
operators is required to prevent the evolution internals from having
large numbers of dispatching functions all over the place.  It is the
responsibility of the caller of `evolution.eigensystem()` to convert the
operators into the new `types.TransformedMatrix` type - in general
usage, this means the base class `System` is responsible.

To achieve this responsibility, while still making subclassing as simple
as possible, various parts of `System` had their access modifiers
changed.  The type `types.TransformedMatrix` is largely internal, so the
methods which return them are the internal `System._hamiltonian()` and
`System._dhamiltonian()`, leaving the call signature of
`System.hamiltonian()` and its derivative untouched.

Of course, subclassing should largely remain unnecessary.