Make integer type-checking less overzealous
Checking for integers as `type(x) == int` is pretty poor style - it
doesn't include subclasses of `int`, for example.

Also, `numpy.int64` and friends don't derive from `int` for some reason,
so we add an explicit check to catch those.

I'm not sure that doing this check is really Python-style - we're
usually meant to just assume it will work, and explode if it doesn't -
but I don't particularly think that's useful in this case.  If the code
were to allow through floats and they didn't cause any exceptions due to
indexing errors or anything, then the resulting Floquet matrix would not
be valid, because we would have violated the assumption that the
Hamiltonian is periodic in the principle frequency.

The mode is probably used as an index at some point, but this shouldn't
be the check - we can imagine a situation where the indexing on the mode
might get removed, and then really bad things could happen unexpectedly.

Better to fail quickly and loudly than slowly and silently.